move `FingerprintRuleJSON` to types module
