+++ src/sentry/grouping/fingerprinting/rules.py
@@ -2,29 +2,18 @@ from __future__ import annotations
 
 import logging
 from collections.abc import Sequence
-from typing import NamedTuple, NotRequired, Self, TypedDict
+from typing import NamedTuple, Self
 
 from sentry.grouping.fingerprinting.matchers import FingerprintMatcher
 from sentry.grouping.fingerprinting.types import (
     FingerprintRuleAttributes,
     FingerprintRuleConfig,
+    FingerprintRuleJSON,
     FingerprintWithAttributes,
 )
 from sentry.grouping.fingerprinting.utils import EventDatastore
 
 logger = logging.getLogger("sentry.events.grouping")
-
-
-# This is just `FingerprintRuleConfig` with an extra `text` entry and with `attributes` required
-# rather than optional. (Unfortunately, you can't overwrite lack of required-ness when subclassing a
-# TypedDict, so we have to create the full type independently.)
-class FingerprintRuleJSON(TypedDict):
-    text: str
-    # Each matcher is a list of [<name of event attribute to match>, <value to match>]
-    matchers: list[list[str]]
-    fingerprint: list[str]
-    attributes: FingerprintRuleAttributes
-    is_builtin: NotRequired[bool]
 
 
 class FingerprintRuleMatch(NamedTuple):
